blog_project_edyoda/requirements.txt
beautifulsoup4==4.9.0
cssselect==1.1.0
Django==2.2.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-tinymce==3.0.1
lxml==4.5.0
markdown2==2.3.8
Pillow==7.1.1
pyquery==1.4.1
pytz==2019.3
readtime==1.1.1
soupsieve==2.0
sqlparse==0.3.1
